Bump to mutwo.core version 2.0.0
It seems as if 'mutwo.common' works for both 'mutwo.core' version:
< 2.0.0 but also >= 2.0.0. We still drop support for pre 2.0.0
'mutwo.core', because otherwise we would need to be sure that we
continue to support both versions, which is due the size of breaking
changes not very easy. Furthermore due to the small amount of mutwo
users, it shouldn't be a problem to just use post 2.0.0 mutwo.